create hash table

+++ 0x1A-hash_tables/0-hash_table_create.c
@@ -7,7 +7,7 @@
 hash_table_t *hash_table_create(unsigned long int size)
 {
 	hash_table_t *hashtable = malloc(sizeof(hash_table_t));
-	unsigned int i; /* to iterate the array */
+	unsigned long int i; /* to iterate the array */
 	/* check if the table was created */
 	if (!hashtable)
 		return (NULL);
